Revert "Cleaning up Mouse / IMouse"
This reverts commit ec2128d1ceb324d8e4b2e69fa5e1fddbba3b2e44.
